+added check for mod perms on clear command

+++ swablu.js
@@ -60,6 +60,11 @@ const command = args.shift().toLowerCase();
     case "clear"://clears a specified amount of messages from chat
 
     const swablu = client.emojis.find("name", "swablu");//creates swablu emoji
+    const modRole = message.guild.roles.find("name", "professors"); //finds mod role
+
+    if (!message.member.roles.has(modRole.id)){ //checks if command user is mod
+      return message.reply("You can't use this command.");
+    }
    
     if(!args[0]) return message.channel.send("Erm, how many? Try again.");//if no amount to delete is given, throws error
     message.channel.bulkDelete((args[0]+1)).then(() => {//does the deleting
